Convert Spinner component to Typescript.
