+++ packages/design-system-docs/src/pages/components/Spinner/_Spinner.docs.scss
@@ -65,7 +65,7 @@ Style guide: components.spinner.background
 
 @react-example Spinner.example.jsx
 
-@react-props Spinner.jsx
+@react-props Spinner.tsx
 
 Style guide: components.spinner.react
 */
